requirements.txt
librosa>=0.10.0
numpy>=1.24.0
pyyaml>=6.0
colorama>=0.4.6
soundfile>=0.12.0
